Split definition of refactoring actions to its own header file
